requirements.txt
typing
python-time
pyyaml
mysql.connector
python-dotenv